requirements.txt
pandas
scikit-learn
azure-storage-blob
python-dotenv
-e .